moves profile provider and project provider to components/providers
